requirements.txt
flask
flask-cors
numpy
pillow
ollama
opencv-python
pdf2image